test(properties): bind the host property suite to the shipped code (#202)
The property suite asserted against private copies of the functions it names,
so it could not observe a regression in any of them. It also inherited the
host timezone, so it only passed in AEST.

## Copies, not code

`edf_properties_test.c` included only `esp_log.h` and `as11_time.h` from
`main/`, and defined its own `clamp_i16`, `snt_missing_for`,
`is_channel_map_valid`, `spool_to_edf` and `format_recording_id`.

Two of the sixteen properties called nothing at all: PROP_09 declared
`int brp_count = 2` and asserted it equalled 2, never reading
`EDF_BRP_SIGNAL_COUNT`; PROP_08 inlined the passthrough branch instead of
exercising the writer.

🔴 And one copy had DRIFTED, which is worse than duplication: the test's
`format_recording_id` called `gmtime_r` where the shipped
`as11_time_format_recording_id` calls `localtime_r`, with the comment "Test
environment uses gmtime_r for UTC baseline". PROP_10 and PROP_16 were green in
every timezone while asserting behaviour the firmware does not have. A copy
that matches is invisible; a copy that has drifted is green and wrong.

Fix: delete the copies and include the shipped headers.
`clamp_i16` / `snt_missing_for` / `is_channel_map_valid` already lived in
`snt_format.h`, and `as11_time_format_recording_id` was already public — only
`spool_to_edf` needed a home, and it moved from `edf_summary.c` into
`edf_data_dict.h` beside the tables it converts for. Both headers are
host-portable (`stdint`/`stdbool`/`limits` only), so no new stubs are needed.

PROP_09 now reads the shipped macros, and additionally checks each table's
length against its own count macro — a macro and its table can disagree.

## Timezone

PROP_01's expectations are pinned to AEST in its comments, but nothing set an
offset: `as11_time_noon_day()` falls back to ESP-local time when none is
stored, and PROP_02 — which parses `+10:00` — runs afterwards. So PROP_01 read
the host clock.

PROP_01 now calls `as11_time_set_offset(10 * 3600, "test")` explicitly, and
`main()` pins `TZ=UTC` for the recording-id properties, which are legitimately
local-time-dependent now that they use the real formatter.

## Result

                          before      after
    Australia/Sydney       37/0       40/0
    UTC                    36/1       40/0
    Europe/Prague          36/1       40/0
    America/New_York       34/3       40/0

Verified by mutation rather than by the suite going green — a suite that was
already green is not evidence:

    A  spool_to_edf rounding -> truncation (the #192 bug)   SURVIVED -> KILLED
    B  EDF_BRP_SIGNAL_COUNT 2 -> 7                          SURVIVED -> KILLED
    C  noon-day previous-day step (control)                 killed   -> killed

C is the control: it was killed before and after, so A and B flipping is the
change and not a difference in how the suite was run.

`as11_time_test` and `as11_events_test` still build and pass. Builds clean
under `-Wall -Wextra`. Net -12 lines.
